backend/pyproject.toml
[project]
name = "streamlog"
version = "0.1.0"
description = "A web-based log management system"
authors = [
    {name = "ShengHongChen",email = "chen36720@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false    